Big figure update.  Fonts are now more consistent.
Many small adjustments as well.  TestOrbit has seen a big change - we now use
plt.scatter() to generate the plot.  If you zoom in, you can see the individual
points and the small scatter from a perfectly circular orbit.

My results for SelfSimilarInfall do not match what's in the paper so I've left
it alone.  This might have something to do with the gravity bug Greg has fixed
in his branch.

Still need to update GravityTest.